chore: renamed error to error-view
