+++ packages/error-view/error-view.js
